moved specialisation to packages to group together by purpose rather than type
