+++ nhais-adaptor/adaptor/outgoing/tests/test_message_adaptor.py
@@ -8,9 +8,9 @@ from fhirclient.models.humanname import HumanName
 from fhirclient.models.address import Address
 from edifact.outgoing.models.name import Name
 from edifact.outgoing.models.address import Address as EdifactAddress
-from edifact.outgoing.models.message import MessageBeginning, MessageTypeBirth
-from edifact.outgoing.models.message_segment_patient_details import MessageSegmentBirthPatientDetails
-from edifact.outgoing.models.message_segment_registration_details import MessageSegmentBirthRegistrationDetails
+from edifact.outgoing.models.message import MessageBeginning
+from edifact.outgoing.models.Birth.message_birth import MessageSegmentBirthPatientDetails, \
+    MessageSegmentBirthRegistrationDetails, MessageTypeBirth
 
 
 class TestMessageAdaptor(unittest.TestCase):
